Coriolis submodule
