+++ src/Coriolis/no_rotation.jl
@@ -1,0 +1,9 @@
+using Oceananigans: AbstractGrid
+
+#####
+##### Functions for non-rotating models
+#####
+
+@inline x_f_cross_U(i, j, k, grid::AbstractGrid{FT}, ::Nothing, U) where FT = zero(FT)
+@inline y_f_cross_U(i, j, k, grid::AbstractGrid{FT}, ::Nothing, U) where FT = zero(FT)
+@inline z_f_cross_U(i, j, k, grid::AbstractGrid{FT}, ::Nothing, U) where FT = zero(FT)
